+++ packages/origin-vue/config.d.ts
@@ -1,0 +1,1 @@
+export { default } from './dist/vite'
