+++ packages/origin-vue/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "@app-research/origin-vue",
+  "version": "0.3.0",
+  "main": "dist/vite/index.js",
+  "types": "dist/vite/index.d.ts",
+  "scripts": {
+    "dev": "npm run build:esm -- -w --incremental",
+    "build": "rm -rf dist && npm run build:esm && npm run build:cjs",
+    "build:esm": "tsc --declaration true",
+    "build:cjs": "tsc --module commonjs --outDir dist/cjs", 
+    "prepublishOnly": "npm run build"
+  },
+  "dependencies": {
+    "@nuxt/devalue": "^2.0.0",
+    "@vueuse/core": "^6.1.0",
+    "@vueuse/head": "^0.5.1",
+    "ipfs-base": "^1.1.0",
+    "pinia": "^2.0.0-rc.6",
+    "vue": "^3.2.2",
+    "vue-router": "^4.0.11"
+  },
+  "files": [
+    "dist",
+    "config.js",
+    "config.d.ts"
+  ],
+  "devDependencies": {
+    "@app-research/eslint-config": "^0.1.2",
+    "@types/node": "^16.7.10",
+    "@vitejs/plugin-vue": "^1.6.0",
+    "@vue/compiler-sfc": "^3.2.8",
+    "@vue/server-renderer": "^3.2.8",
+    "autoprefixer": "^10.3.4",
+    "postcss-nested": "^5.0.6",
+    "typescript": "^4.4.2",
+    "unplugin-vue-components": "^0.14.11",
+    "vite": "^2.5.3",
+    "vite-plugin-pages": "^0.18.0",
+    "vite-plugin-vue-layouts": "^0.3.1",
+    "vite-ssg": "^0.14.7",
+    "vue-tsc": "^0.3.0"
+  }
+}
