+++ packages/origin-vue/src/origin/index.ts
@@ -1,0 +1,1 @@
+export { defineConfig, omit } from './origin.config'
