+++ packages/origin-vue/src/vite/index.ts
@@ -1,0 +1,1 @@
+export { default } from './vite.config'
